Bug 462873 - Change DMR to use segments
Started to implement model migration for mapping references.
There's a quite strange bug:
After the migration, the validation fails for the EClass set as the
mappedClass in the mapping segment. During the validation the instance
type of the the EClass's instance is EClass (as it should be). But for
some reason the reference type of the mappedClass EReference is
EFactory.
The validation fails in
org.eclipse.emf.edapt.spi.migration.util.MigrationValidator
#validate_validType(ReferenceSlot) (line573)

Otherwise the migration implementation is probably complete.

Change-Id: I9386bfc3413bc834b064bd93603133990a1ef752
Signed-off-by: Lucas Koehler <l.koehler93@gmail.com>