Bug 542673 - [Segments] Provide migration with persistence from legacy DMRs to segments
Legacy Dmr to segments migration in the ViewEditorPart:
If the segment tooling is enabled, the view editor part now checks
whether an opened view model contains legacy dmrs. If it does, the user
is queried whether (s)he wants to migrate to segment based dmrs.
The check is done after the ViewEditorPart has checked for necessary
Edapt migrations. If an Edapt migrations was necessary and the user
declined or the migration failed, the editor will not check for legacy
dmr migration.

Further implementations:
* Create new bundle org.eclipse.emfforms.ide.view.segments for segment
related ide/tooling services
* Create corresponding test fragment for the new bundle
* Move ToolingModeUtil to the new bundle
* Add the DmrToSegmentsMigrator service which allows to check whether a
view model still contains legacy dmrs and to migrate them to segment
based dmrs
* Add tests for the DmrToSegmentsMigrator implementation
* Adapt the TableDmrSegmentGenerator to generating new child DMRs for
the generated multi segment instead of copying the existing column DMRs.

Change-Id: I1eb412d09ebbbc6862f16207b23c4e511d690cf7
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>