Bug 527686 - Pull-out common master-detail code
Factor out cache of detail renderings from the tree
master-detail composite for reuse. Use a StackLayout
to avoid reparenting controls into and out of a secret
“limbo” shell.

Implement support for caching in the master-detail
tree renderer, table renderer, and grid renderer.

Fix numerous problems with management of child
contexts in conjunction with replacement of the
domain model object (required for context reuse)
and internal disposal/reinitialization of contexts.

Fix some renderers that didn’t properly support
replacement of the domain model object and/or
caching of controls.

Fix validation service and control mapper to
properly handle reuse of child contexts.

Change-Id: I3b3979bf822f4a34509fdeb3677038fc5a72c10e
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>