Bug 548330 - Fix column config reference service for segments
Adapt the ColumnConfigurationDMRRendererReferenceService to also handle
tables based on a DMR with a multi segment.
If there are segments in the dmr, the reference service now searches for
the multi segment to get its column dmrs.
If there are no segments, the ref. service tries to convert the dmr to a
table dmr.

Change-Id: Ibca1ac57212fd82f83d6feb4fe83573a07bb7c4b
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>