Bug 549270 - Readonly controls should not show set/unset button
If the link control's VControl is effectively readonly, the add, create
and link, and delete buttons are now hidden.
This change does NOT hide the (un)set button for an unsettable link
control because this is a more universal issue which needs fixing in a
super class.

Additionally, cleaned up warnings in the link control unit tests.

Change-Id: I7869cd9871836ac342e1d2e203d2b01528fbdb2a
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>