Bug 499923 - Link Control should not show buttons when read-only
If the link control's VControl is effectively readonly, the add, create
and link, and delete buttons are now hidden.
This change does NOT hide the (un)set button for an unsettable link
control because this is a more universal issue which needs fixing in a
super class.

Additionally, cleaned up warnings in the link control unit tests.

Change-Id: I88f4cbb7de565eae7a20b46b555541a01e46dd42
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>