Bug 541190 - Generic Editor should support more input sources than files
Adapt the Generic Editor to handle opening files from further inputs
besides files in the workspace. For instance, the Generic Editor now
supports opening files from a local or Git history revision.
If an opened editor input does not allow to persist the loaded data
(e.g. in the history revision case), the editor is opened as read-only.
Concretely, this has the following implications:
* Drag and drop is disabled for the tree
* The tree's context menu is disabled
* The editor doesn't show toolbar actions
* All detail views are shown as read-only but the user can still select
different objects in the tree to view their detail views

To achieve the foregoing behavior, the TreeViewerCustomization was
extended by a read-only configuration which can be queried from a
customization implementation.
The DefaultTreeViewerCustomization was adapted to this by not returning
any menu and stating that drag and drop is disabled whenever read-only
is set to true.

Change-Id: I76972d2399c05fdd908ed45e5c713da7842a3fa9
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>