Bug 546519 - Validation Service cannot handle most nested DMRs
The notifyAdd method of the validation service's view model change
listener now only validates the EObjects of an added DMR if the DMR is
directly contained in a VControl.

This is enough because only Settings rendered by controls are relevant
for validation.

With this, nested DMRs no longer get unnecessarily resolved by the view
model change listener and, thus, the log is no longer spammed with
meaningless DatabindingFailedExceptions.

Change-Id: I33c57dd35543fc810829589cf9b8c761e77c8fde
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>