Bug 549634 - View model editor should support opening files from history
The view model editor can now open view models from a history revision.
If this is the case, the view model is copied to a temporary file to
allow migration before showing it in the editor.
View models opened from the history are shown as readonly. For this the
tree master detail renderer was adapted to NOT offer drag and drop or a
context menu if the tree is configured as readonly

Also added a RCPTT test testing the opening from history inluding
migration.

Change-Id: Ifbe35d0e7470fa1d9bbc9de1999f6ca7c9c20092
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>