Bug 553181 - Allow customization of validation icon and color calculation
Add the new SWTValidationUiService which encapsulates getting validation
icons and colors. This allows replacing the logic by registering a
custom service implementation with higher priority.
The default implementation delegates to the SWTValidationHelper.
Added three new protected final methods to the
AbstractControlSWTRenderer which use the new service. Deprecated the old
methods which request a severity as an argument.

Copied the SWTValidationHelper to bundle emfforms.swt.core and
deprecated the original version in emf.ecp.edit.swt.

Added unit tests for the new service and added unit tests for
AbstractControlSWTRenderer.

Change-Id: Ie284e9a34db6799a975850252ab5e879f8ea431b
Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>